pyproject.toml
[project]
name = "ssu-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "feedparser>=6.0.11",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langchain-community>=0.3.27",
    "pymilvus>=2.6.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
    "openai>=1.0.0",
]
